pyproject.toml
[project]
name = "open-data-product-hamburg-points-of-interest-source-aligned"
version = "0.1.0"
description = "Source-aligned data product providing Hamburg points-of-interest data"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "open-data-product-python-lib",
]

[tool.uv.sources]
open-data-product-python-lib = { git = "https://github.com/open-data-product/open-data-product-python-lib.git" }
